Separate files for different resource tests
